pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "LucasGabrielAntonete", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

